requirements.txt
# CDW Experiments - Requirements
# ==============================

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.5.0

# Diffusion models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Image quality metrics
lpips>=0.1.4
torchmetrics>=1.0.0
scikit-image>=0.21.0

# Machine learning utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration and serialization
pyyaml>=6.0
omegaconf>=2.3.0

# Plotting and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0
